Improve Flask tests
* Build a completely fresh Flask application in the context manager, rather than try unpatch what `ScoutApm` does
* Add assertions on the `tracked_requests` fixture around what we expect to be tracked
